Add files via upload
Used Power BI to create a dashboard to show sales, profits and costs of different areas of United States. The dashboard was used to analyse E-commerce sales in different areas of US. Maximum sales turned out to be in technology’s office machines with value of $318,170. New York and California turned out as the most profitable areas for business.